+++ src/dApps/AccessChannels/Origin.ts
@@ -1,0 +1,3 @@
+export interface Origin {
+  origin: string;
+}
