+++ src/dApps/AccessChannels/backgroundAccessChannels.ts
@@ -1,8 +1,9 @@
 import { PopupChannel } from '../../channels/base/PopupChannel/PopupChannel';
 
 import { contentAccessChannel } from './contentAccessChannel';
-import { getAuthorizedOrigin, Origin } from './getAuthorizedOrigin';
+import { getAuthorizedOrigin} from './getAuthorizedOrigin';
 import { AccessInput } from './types';
+import { Origin } from './Origin';
 
 export const backgroundAccessChannel = new PopupChannel<
   AccessInput & Origin,
