+++ src/views/AuthorizeDApp/AuthorizeDApp.tsx
@@ -6,7 +6,7 @@ import * as styles from './AuthorizeDApp.module.css';
 import { usePopupData } from '../../utilities/popups/usePopupData';
 import { backgroundAccessChannel } from '../../dApps/AccessChannels/backgroundAccessChannels';
 import { AccessInput } from '../../dApps/AccessChannels/types';
-import { Origin } from '../../dApps/AccessChannels/getAuthorizedOrigin';
+import { Origin } from '../../dApps/AccessChannels/Origin';
 
 export function AuthorizeDApp(): JSX.Element {
   const t = browser.i18n.getMessage;
